coreAPI/requirements.txt
anyio==3.5.0
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
coverage==7.2.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
gTTS==2.0.3
gTTS-token==1.1.3
h11==0.13.0
html5lib==1.1
httpcore==0.14.5
httplib2==0.20.2
httpx==0.22.0
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.1.2
jeepney==0.7.1
kazam==1.4.5
keyring==23.5.0
langdetect==1.0.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.1.1
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2022.11.25
platformdirs==2.5.4
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rfc3986==1.5.0
rich==11.2.0
scour==0.38.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.6.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.7
virtualenv-clone==0.5.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
